backend/requirements.txt
asgiref==3.6.0
accessify==0.3.1
Django==4.1.5
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-dotenv
sqlparse==0.4.3
requests==2.28.2
Pillow==9.3.0
fpdf2==2.7.4